Conflicts:
	CBV.h

+++ CBV.cpp
@@ -640,9 +640,9 @@ void CBV::Concat(const CBV& bv1,const CBV& bv2)
 
 //****************************** Protected function ***********************************//
 //---------------------------------------------------------- LeftShiftInPlace(int nShift)
-void CBV::LeftShiftInPlace(int nShift)
+void CBV::LeftShiftInPlace(size_t nShift)
 {  // LeftShift bool vector  -- assume that 'this' is a new CBV object
- int i,j,l_bit,r_bit,nNewLenByte,nNewLenBit,AdrBeg;
+ size_t i,j,l_bit,r_bit,nNewLenByte,nNewLenBit,AdrBeg;
 
  ASSERT(nShift >= 0);
  nNewLenBit = m_nBitLength - nShift;
@@ -665,9 +665,9 @@ void CBV::LeftShiftInPlace(int nShift)
 }
 
 //--------------------------------------------------------- RightShiftInPlace(int nShift)
-void CBV::RightShiftInPlace(int nShift)
+void CBV::RightShiftInPlace(size_t nShift)
 {  // RightShift bool vector  -- assume that 'this' is a new CBV object
- int i,j,l_bit,r_bit,AdrBeg;
+ size_t i,j,l_bit,r_bit,AdrBeg;
  BYTE *work;
  ASSERT(nShift >= 0);
  ASSERT((m_nBitLength - nShift) >= 0);
@@ -691,32 +691,32 @@ void CBV::RightShiftInPlace(int nShift)
 }
 
 //--------------------------------------------- DizInPlace(const BYTE* Vect1,int BitLen1)
-void CBV::DizInPlace(const BYTE* Vect1,int BitLen1)
-{int i;
+void CBV::DizInPlace(const BYTE* Vect1,size_t BitLen1)
+{size_t i;
  ASSERT(BitLen1 >= 0);
  ASSERT(BitLen1 == m_nBitLength);
  for (i=0; i<m_nByteLength; i++) m_bVect[i] = m_bVect[i] | Vect1[i];
 }
 
 //--------------------------------------------- ConInPlace(const BYTE* Vect1,int BitLen1)
-void CBV::ConInPlace(const BYTE* Vect1,int BitLen1)
-{int i;
+void CBV::ConInPlace(const BYTE* Vect1,size_t BitLen1)
+{size_t i;
  ASSERT(BitLen1 >= 0);
  ASSERT(BitLen1 == m_nBitLength);
  for (i=0; i<m_nByteLength; i++) m_bVect[i] = m_bVect[i] & Vect1[i];
 }
 
 //-------------------------------------------- Add2InPlace(const BYTE* Vect1,int BitLen1)
-void CBV::Add2InPlace(const BYTE* Vect1,int BitLen1)
-{int i;
+void CBV::Add2InPlace(const BYTE* Vect1,size_t BitLen1)
+{size_t i;
  ASSERT(BitLen1 >= 0);
  ASSERT(BitLen1 == m_nBitLength);
  for (i=0; i<m_nByteLength; i++) m_bVect[i] = m_bVect[i] ^ Vect1[i];
 }
 
 //------------------------------------------ ConNotInPlace(const BYTE* Vect1,int BitLen1)
-void CBV::ConNotInPlace(const BYTE* Vect1,int BitLen1)
-{int i;
+void CBV::ConNotInPlace(const BYTE* Vect1,size_t BitLen1)
+{size_t i;
  ASSERT(BitLen1 >= 0);
  ASSERT(BitLen1 == m_nBitLength);
  for (i=0; i<m_nByteLength; i++) m_bVect[i] = m_bVect[i] & ~Vect1[i];
@@ -726,11 +726,11 @@ void CBV::ConNotInPlace(const BYTE* Vect1,int BitLen1)
 //*************************************************************************************//
 
 //-------------------------------------------------------------- operator <<=(int nShift)
-const CBV&  CBV::operator <<=(int nShift)
+const CBV&  CBV::operator <<=(size_t nShift)
 { LeftShiftInPlace(nShift);  return *this; }
 
 //-------------------------------------------------------------- operator >>=(int nShift)
-const CBV& CBV::operator >>=(int nShift)
+const CBV& CBV::operator >>=(size_t nShift)
 { RightShiftInPlace(nShift); return *this;}
 
 //-------------------------------------------------------- operator |=(const CBV& bv1)
@@ -769,9 +769,9 @@ const CBV& CBV::operator -=(const BYTE* pbt)
 /////////////////////////////////// Logic operations ////////////////////////////////////
 
 //****************************** Protected function ***********************************//
-//--------------------------- LeftShiftVect(const BYTE* SrcVect, int nSrcLen, int nShift)
-void CBV::LeftShiftVect(const BYTE* SrcVect, int nSrcLen, int nShift)
-{int j,i,l_bit,r_bit,nNewLenByte,nNewLenBit,nSrcByte,AdrBeg;
+//--------------------------- LeftShiftVect(const BYTE* SrcVect, size_t nSrcLen, size_t nShift)
+void CBV::LeftShiftVect(const BYTE* SrcVect, size_t nSrcLen, size_t nShift)
+{size_t j,i,l_bit,r_bit,nNewLenByte,nNewLenBit,nSrcByte,AdrBeg;
  ASSERT(nShift >= 0);
  nNewLenBit = nSrcLen - nShift;
  ASSERT(nNewLenBit >= 0);
@@ -795,9 +795,9 @@ void CBV::LeftShiftVect(const BYTE* SrcVect, int nSrcLen, int nShift)
  memset(m_bVect+nNewLenByte, 0, i);
 }
 
-//-------------------------- RightShiftVect(const BYTE* SrcVect, int nSrcLen, int nShift)
-void CBV::RightShiftVect(const BYTE* SrcVect, int nSrcLen, int nShift)
-{int i,j,l_bit,r_bit,nSrcByte,AdrBeg;
+//-------------------------- RightShiftVect(const BYTE* SrcVect, size_t nSrcLen, size_t nShift)
+void CBV::RightShiftVect(const BYTE* SrcVect, size_t nSrcLen, size_t nShift)
+{size_t i,j,l_bit,r_bit,nSrcByte,AdrBeg;
  ASSERT(nShift >= 0);
  ASSERT((nSrcLen - nShift) >= 0);
  nSrcByte = LEN_BYTE(nSrcLen);
@@ -818,9 +818,9 @@ void CBV::RightShiftVect(const BYTE* SrcVect, int nSrcLen, int nShift)
    m_bVect[nSrcByte-1] = (m_bVect[nSrcByte-1] >> (S_1 - i)) << (S_1 - i);
 }
 
-//---------------------- Diz(const BYTE* Vect1,int BitLen1,const BYTE* Vect2,int BitLen2)
-void CBV::Diz(const BYTE* Vect1,int BitLen1,const BYTE* Vect2,int BitLen2)
-{int i,LenByte;
+//---------------------- Diz(const BYTE* Vect1,size_t BitLen1,const BYTE* Vect2,size_t BitLen2)
+void CBV::Diz(const BYTE* Vect1,size_t BitLen1,const BYTE* Vect2,size_t BitLen2)
+{size_t i,LenByte;
  ASSERT(BitLen1 == BitLen2);
  ASSERT(BitLen1 >= 0);
  LenByte = LEN_BYTE(BitLen1);
@@ -829,9 +829,9 @@ void CBV::Diz(const BYTE* Vect1,int BitLen1,const BYTE* Vect2,int BitLen2)
  for (i=0; i<LenByte; i++) m_bVect[i] = Vect1[i] | Vect2[i];
 }
 
-//---------------------- Con(const BYTE* Vect1,int BitLen1,const BYTE* Vect2,int BitLen2)
-void CBV::Con(const BYTE* Vect1,int BitLen1,const BYTE* Vect2,int BitLen2)
-{int i,LenByte;
+//---------------------- Con(const BYTE* Vect1,size_t BitLen1,const BYTE* Vect2,size_t BitLen2)
+void CBV::Con(const BYTE* Vect1,size_t BitLen1,const BYTE* Vect2,size_t BitLen2)
+{size_t i,LenByte;
  ASSERT(BitLen1 == BitLen2);
  ASSERT(BitLen1 >= 0);
  LenByte = LEN_BYTE(BitLen1);
@@ -840,9 +840,9 @@ void CBV::Con(const BYTE* Vect1,int BitLen1,const BYTE* Vect2,int BitLen2)
  for (i=0; i<LenByte; i++) m_bVect[i] = Vect1[i] & Vect2[i];
 }
 
-//--------------------- Add2(const BYTE* Vect1,int BitLen1,const BYTE* Vect2,int BitLen2)
-void CBV::Add2(const BYTE* Vect1,int BitLen1,const BYTE* Vect2,int BitLen2)
-{int i,LenByte;
+//--------------------- Add2(const BYTE* Vect1,size_t BitLen1,const BYTE* Vect2,size_t BitLen2)
+void CBV::Add2(const BYTE* Vect1,size_t BitLen1,const BYTE* Vect2,size_t BitLen2)
+{size_t i,LenByte;
  ASSERT(BitLen1 == BitLen2);
  ASSERT(BitLen1 >= 0);
  LenByte = LEN_BYTE(BitLen1);
@@ -851,9 +851,9 @@ void CBV::Add2(const BYTE* Vect1,int BitLen1,const BYTE* Vect2,int BitLen2)
  for (i=0; i<LenByte; i++) m_bVect[i] = Vect1[i] ^ Vect2[i];
 }
 
-//------------------- ConNot(const BYTE* Vect1,int BitLen1,const BYTE* Vect2,int BitLen2)
-void CBV::ConNot(const BYTE* Vect1,int BitLen1,const BYTE* Vect2,int BitLen2)
-{int i,LenByte;
+//------------------- ConNot(const BYTE* Vect1,size_t BitLen1,const BYTE* Vect2,size_t BitLen2)
+void CBV::ConNot(const BYTE* Vect1,size_t BitLen1,const BYTE* Vect2,size_t BitLen2)
+{size_t i,LenByte;
  ASSERT(BitLen1 == BitLen2);
  ASSERT(BitLen1 >= 0);
  LenByte = LEN_BYTE(BitLen1);
@@ -862,9 +862,9 @@ void CBV::ConNot(const BYTE* Vect1,int BitLen1,const BYTE* Vect2,int BitLen2)
  for (i=0; i<LenByte; i++) m_bVect[i] = Vect1[i] & ~Vect2[i];
 }
 
-//---------------------------------------------------- Not(const BYTE* Vect1,int BitLen1)
-void CBV::Not(const BYTE* Vect1,int BitLen1)
-{int i,LenByte;
+//---------------------------------------------------- Not(const BYTE* Vect1,size_t BitLen1)
+void CBV::Not(const BYTE* Vect1,size_t BitLen1)
+{size_t i,LenByte;
  ASSERT(BitLen1 >= 0);
  LenByte = LEN_BYTE(BitLen1);
  AllocBuffer(LenByte);
@@ -875,14 +875,14 @@ void CBV::Not(const BYTE* Vect1,int BitLen1)
 }
 //*************************************************************************************//
 
-//------------------------------------- operator <<(const CBV& bv1, int nShift)
-STD(CBV) operator <<(const CBV& bv1, int nShift)
+//------------------------------------- operator <<(const CBV& bv1, size_t nShift)
+STD(CBV) operator <<(const CBV& bv1, size_t nShift)
 { CBV s;
  s.LeftShiftVect(bv1.m_bVect,bv1.m_nBitLength,nShift);  return s;
 }
 
-//------------------------------------- operator >>(const CBV& bv1, int nShift)
-STD(CBV) operator >>(const CBV& bv1, int nShift)
+//------------------------------------- operator >>(const CBV& bv1, size_t nShift)
+STD(CBV) operator >>(const CBV& bv1, size_t nShift)
 { CBV s;
  s.RightShiftVect(bv1.m_bVect,bv1.m_nBitLength,nShift);  return s;
 }
@@ -980,16 +980,16 @@ return s;
 
 //-------------------------------------------------- Invert(const BYTE* pbt)
 void CBV:: Invert(const BYTE* pbt)
-{ int i;
+{ size_t i;
  for (i=0; i<m_nByteLength; i++) m_bVect[i] = ~pbt[i];
  i = ADR_BIT(m_nBitLength);
  if (i)  m_bVect[m_nByteLength-1] = (m_bVect[m_nByteLength-1] >> (S_1-i)) << (S_1-i);
 }
 
-//------------------------------------------------------------- LoopLeftShift(int nShift)
-void CBV::LoopLeftShift(int nShift)
+//------------------------------------------------------------- LoopLeftShift(size_t nShift)
+void CBV::LoopLeftShift(size_t nShift)
 {  // LeftShift bool vector  -- assume that 'this' is a new CBV object
- int i,j,l_bit,r_bit,nNewLenByte,nNewLenBit,nEndLenByte;
+ size_t i,j,l_bit,r_bit,nNewLenByte,nNewLenBit,nEndLenByte;
  BYTE *work;
  ASSERT(nShift >= 0);
  nNewLenBit = m_nBitLength - nShift;
@@ -1025,10 +1025,10 @@ void CBV::LoopLeftShift(int nShift)
  delete [] work;
 }
 
-//------------------------------------------------------------ LoopRightShift(int nShift)
-void CBV::LoopRightShift(int nShift)
+//------------------------------------------------------------ LoopRightShift(size_t nShift)
+void CBV::LoopRightShift(size_t nShift)
 {  // LeftShift bool vector  -- assume that 'this' is a new CBV object
- int i,j,l_bit,r_bit,AdrBeg,AdrEnd,BitEnd;
+ size_t i,j,l_bit,r_bit,AdrBeg,AdrEnd,BitEnd;
  BYTE *work;
  ASSERT(nShift >= 0);
  ASSERT((m_nBitLength - nShift) >= 0);
@@ -1071,10 +1071,10 @@ int CBV::CountBit() const
  return one;
 }
 
-//-------------------------------------------------------------------- LeftOne(int nNext)
-int CBV::LeftOne(int nNext) const
-{int i,j,pos = 0;
- int k;                                //new 24.01.00
+//-------------------------------------------------------------------- LeftOne(ptrdiff_t nNext)
+ptrdiff_t CBV::LeftOne(ptrdiff_t nNext) const
+{ptrdiff_t i,j,pos = 0;
+ size_t k;                                //new 24.01.00
  BYTE ch;
 // ASSERT((m_nBitLength - nNext) >= 0);
  ASSERT((m_nBitLength - nNext) > 0);   //new 24.01.00
@@ -1099,8 +1099,8 @@ int CBV::LeftOne(int nNext) const
 }
 
 //--------------------------------------------------------------------- LeftOne(BYTE& bt)
-int CBV::LeftOne(BYTE& bt) const
-{int i,pos;
+ptrdiff_t CBV::LeftOne(BYTE& bt) const
+{ptrdiff_t i,pos;
  for (i=0;i<m_nByteLength;i++)
    if (m_bVect[i]!=0) {
      pos=0;
